added function to strip data frame and clean list
